+++ src/house.h
@@ -75,7 +75,7 @@ class Door : public Item
 		virtual bool serializeAttr(PropWriteStream& propWriteStream) const;
 
 		void setDoorId(uint32_t _doorId) {
-			setIntAttr(ATTR_ITEM_DOORID, (uint32_t)_doorId);
+			setIntAttr(ATTR_ITEM_DOORID, _doorId);
 		}
 		uint32_t getDoorId() const {
 			return getIntAttr(ATTR_ITEM_DOORID);
